+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Fixed:
 - Fixed incorrect top-bottom cannon cast merger fluid capacity
 - Fixed cannon builder assembly and layering collision
 - Fixed white seat flickering when changing autocannon seating when Flywheel is on
+- Fixed tracer spacing of regular Autocannon Ammo Container
 
 ## [5.10.2] - 2026-01-17